requirements.txt
python-decouple
requests~=2.25.1
pillow~=8.2.0